# Refine the activation type getting in the LSTM operator to speed.